+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.8.0]
+- Improve course administration course filter
+- Fix older ILIAS 5.4 versions
+
 ## [2.7.0]
 - ILIAS 6 support
 - Remove ILIAS 5.3 support